perf: Pre-bake Qwen2-0.5B model and switch to BFloat16
Optimizations to reduce SLM service startup time from >10min to ~2-3min:

1. Pre-download model during Docker build:
   - Qwen2-0.5B-Instruct model (~1GB) now included in container
   - Eliminates 3-5min HuggingFace download on cold start
   - Model cached in /root/.cache/huggingface inside container

2. Switch from float32 to bfloat16:
   - Reduces memory footprint by 50% (2GB → 1GB)
   - Faster CPU operations with BF16
   - Native precision for modern transformers

Expected startup time: 2-3 minutes (was >10 minutes)
Container size increase: ~+2GB
Cost: No change (~$151/month)

+++ dpr_rc/slm_service.py
@@ -236,7 +236,7 @@ def load_model():
 
         _model = AutoModelForCausalLM.from_pretrained(
             SLM_MODEL,
-            torch_dtype=torch.float32 if DEVICE == "cpu" else torch.float16,
+            torch_dtype=torch.bfloat16 if DEVICE == "cpu" else torch.float16,
             device_map="auto" if DEVICE == "cuda" else None,
             trust_remote_code=True
         )
